backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.10.0
